==> ordinal_encoding/1.Scaling and  Transforming.ipynb <==
import numpy as np 
import pandas as pd 
from pandas import DataFrame
import os
import joblib
import sklearn
from sklearn.svm import SVR
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import StandardScaler, OrdinalEncoder

# Import plotting libraries
import seaborn as sns
import matplotlib 
from matplotlib import pyplot as plt

# Set larger fontsize for all plots
matplotlib.rcParams.update({'font.size': 30})
from IPython.display import clear_output

# Getting dataset
df = pd.read_csv('../raw data.csv')
df

#Checks if there are any columns with empty cells
df.isna().sum()

# Scaling and Transforming

df_input = df.drop(columns =['DOI','method','metal_source','size_nm','Absorbance','Ellip_mdeg','g_factor*10^4.1','g_factor*10^4'], inplace = False, axis = 1)
df_output = df[['size_nm','g_factor*10^4.1']]

#Checks the column names, and ensures that they do not have any leading or trailing spaces
df_input.columns = df_input.columns.str.strip()
df_output.columns = df_output.columns.str.strip()

df_input.info()

#change temp and time from integer to float
df_input['temp_c'] = df_input['temp_c'].astype(float)
df_input['time_min'] = df_input['time_min'].astype(float)
df_input.info()

#Initializes 2 lists to contain all the numerical and categorical input columns
input_num_cols = [col for col in df_input.columns if df[col].dtypes !='O']
input_cat_cols = [col for col in df_input.columns if df[col].dtypes =='O']
print(input_num_cols, input_cat_cols)

#Initializes the ColumnTransformer object, and specifies what it will do with a dataframe
#scaling numerical columns
#onehotencoder creates a binary column for each categorical entry
ct = ColumnTransformer([
    ('step1', StandardScaler(), input_num_cols),
    ('step2', OrdinalEncoder( handle_unknown='error'), input_cat_cols)
], remainder = 'passthrough')
#what are the transformers in ct
ct.transformers

#Uses the ColumnTransformer object to modify the input columns
df_input_scaled_encoded = pd.DataFrame(ct.fit_transform(df_input))
df_input_scaled_encoded

#Number of numerical columns
len(input_num_cols)

#renaming columns
for i in range(len(input_num_cols)):
    df_input_scaled_encoded.rename(columns={df_input_scaled_encoded.columns[i]: input_num_cols[i]}, inplace = True)

df_input_scaled_encoded.columns

#appends output columns
df_scaled_encoded = pd.concat([df_input_scaled_encoded, df_output], axis = 1)
df_scaled_encoded

df_scaled_encoded.to_csv('dataset_scaled.csv')

# Making separate outputs dataset

#Saves into a list the row indexes to drop for size dataset
total_row_num = len(df_scaled_encoded)
drop_list_size =[]
for row_i in range(total_row_num):
    if df_scaled_encoded['size_nm'].values[row_i] == 'None':
        drop_list_size.append(row_i)
    
#number of entries
print(total_row_num-len(drop_list_size))

#Drops rows that don't have size output
df_size_scaled_encoded = df_scaled_encoded.drop(drop_list_size)
df_size_scaled_encoded.to_csv('dataset_scaled_size.csv')
